Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
anthropic>=0.30.0
python-dotenv==1.0.0
boto3==1.34.34